requirements.txt
chromadb
transformers
python-telegram-bot
